feat: eligibliity check on after personal data
